requirements.txt
Flask==3.1.*
Flask-WTF==1.2.*
Flask-Cors==4.0.*
python-dotenv==1.0.*
Werkzeug==3.1.*
python-slugify==8.0.*
gunicorn==21.2.*
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv
requests